crud logic is done for all element.
crud is compeleted for all elements and there are only some logical trap that need to be fixed for the thing to run. but we are going to start fixiating on the visual aspect of the program now. I will get the assistents form a friend to help design the generale layout and then add some functions behind the new layout and add features, like the abillity to change color and font/font-size and many other things. but right now make it pretty and readable