added some graphical elemets
the look is sorta done and now i have to work on the drag and drop feature of the cards. Need to make a trigger or storage procedur in the slq database, so thatI can update the view and data when the user d&d something anywhere in the view. Might not be posible to do this at home, but I will at least be able to make some progress on it and learn the ins and outs of html and vue. hvis det ikke ser ud til at kunne gøres kan jeg lave nogle små pile som man kan diagere med.